requirements.txt
mcp>=1.26.0
docker>=7.0.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
stix2>=3.0.0
